add to basket corrected

+++ database/migrations/2024_08_24_045811_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->decimal('total_price');
+            $table->bigInteger('total_price');
             $table->string('status')->default('pending');
             $table->softDeletes();
             $table->timestamps();
